requirements.txt
pymupdf
sentence-transformers
faiss-cpu
openai
langchain
python-dotenv
streamlit
groq